LPS-65693 Pull out validation logic as a Component for reuse
