+++ modules/core/portal-lpkg-deployer/src/main/java/com/liferay/portal/lpkg/deployer/internal/DefaultLPKGVerifier.java
@@ -14,18 +14,12 @@
 
 package com.liferay.portal.lpkg.deployer.internal;
 
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
-
 import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayOutputStream;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.lpkg.deployer.LPKGVerifier;
 import com.liferay.portal.lpkg.deployer.LPKGVerifyException;
-import com.liferay.portal.target.platform.indexer.IndexValidator;
-import com.liferay.portal.target.platform.indexer.IndexValidatorFactory;
 import com.liferay.portal.target.platform.indexer.Indexer;
 import com.liferay.portal.target.platform.indexer.IndexerFactory;
 import com.liferay.portal.util.PropsValues;
@@ -34,16 +28,12 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 
-import java.net.URI;
-
-import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Properties;
 import java.util.zip.ZipEntry;
@@ -168,47 +158,8 @@ public class DefaultLPKGVerifier implements LPKGVerifier {
 				_log.info("Wrote index " + indexFilePath);
 			}
 
-			IndexValidator indexValidator = _indexValidatorFactory.create(
-				_getTargetPlatformIndexURIs());
-
-			long start = System.currentTimeMillis();
-
-			try {
-				List<String> messages = indexValidator.validate(
-					Collections.singletonList(indexFilePath.toUri()));
-
-				if (!messages.isEmpty()) {
-					Files.delete(indexFilePath);
-
-					StringBundler sb = new StringBundler(
-						(messages.size() * 3) + 1);
-
-					sb.append("LPKG validation failed with {");
-
-					for (String message : messages) {
-						sb.append("[");
-						sb.append(message);
-						sb.append("], ");
-					}
-
-					sb.setIndex(sb.index() - 1);
-
-					sb.append("]}");
-
-					throw new LPKGVerifyException(sb.toString());
-				}
-			}
-			finally {
-				if (_log.isInfoEnabled()) {
-					long duration = System.currentTimeMillis() - start;
-
-					_log.info(
-						String.format(
-							"LPKG validation time %02d:%02ds",
-							MILLISECONDS.toMinutes(duration),
-							MILLISECONDS.toSeconds(duration % Time.MINUTE)));
-				}
-			}
+			_lpkgIndexValidator.validate(
+				Collections.singletonList(indexFilePath.toUri()));
 		}
 		catch (LPKGVerifyException lpkgve) {
 			throw lpkgve;
@@ -216,27 +167,6 @@ public class DefaultLPKGVerifier implements LPKGVerifier {
 		catch (Exception e) {
 			throw new LPKGVerifyException(e);
 		}
-	}
-
-	private List<URI> _getTargetPlatformIndexURIs() throws IOException {
-		List<URI> uris = new ArrayList<>();
-
-		try (DirectoryStream<Path> directoryStream = Files.newDirectoryStream(
-				Paths.get(
-					PropsValues.MODULE_FRAMEWORK_BASE_DIR,
-					Indexer.DIR_NAME_TARGET_PLATFORM),
-				"*.xml")) {
-
-			Iterator<Path> iterator = directoryStream.iterator();
-
-			while (iterator.hasNext()) {
-				Path path = iterator.next();
-
-				uris.add(path.toUri());
-			}
-		}
-
-		return uris;
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
@@ -248,6 +178,6 @@ public class DefaultLPKGVerifier implements LPKGVerifier {
 	private IndexerFactory _indexerFactory;
 
 	@Reference
-	private IndexValidatorFactory _indexValidatorFactory;
+	private LPKGIndexValidator _lpkgIndexValidator;
 
 }
